### 2019-07-26 #40 @whedon
[PRE REVIEW]: GlobalSearchRegression.jl: Building bridges between Machine Learning and Econometrics in Fat-Data scenarios
**Submitting author:** @dpanigo (<a href="http://orcid.org/0000-0003-2468-1632">Demian Tupac PANIGO</a>)
**Repository:** <a href="https://github.com/ParallelGSReg/JuliaCon2019" target ="_blank">https://github.com/ParallelGSReg/JuliaCon2019</a>

Thanks for submitting your paper to JuliaCon @dpanigo. **Currently, there isn't a JuliaCon editor assigned** to your paper.

@dpanigo if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
