### 2019-07-26 120e98e
first round of fixes
