### 2019-07-26 84bf015
Update paper.tex
