+++ paper/paper.tex
@@ -12,376 +12,6 @@ The aim of this paper is twofold. The first one is to describe a novel research-
 
 \end{abstract}
 
-\section{Introduction}
-While allowing for better and more accurate analysis (and forecasts), fat-data availability generates a revival of Bellman’s  -and newer- high-dimensionality concerns:
-\begin{enumerate}
-    \item Exponential increase in required information and execution time;
-    \item Overfitting risk ; 
-    \item Less-informative euclidean distances; and
-    \item Unfeasible estimators. 
-\end{enumerate}
-
-Notwithstanding, the advantage of having thousands/millions of features to deal with complex phenomena stimulates an unprecedented number of methodological -and technological- improvements to manage the ‘curse of dimensionality’. \vskip 6pt
-
-In Economics, this process has a dual approach with machine-learning (ML) and econometric (EC) algorithms emerging for different purposes: the former for prediction/forecasts (focusing on $\hat{y}$) and the latter for estimation/causal inference (interested in $\hat{\beta}$). Alternatively, the same distinction can be expressed in Diebold’s terms as non-causal vs causal prediction, where ML algorithms are designed to reduce prediction sampling-risks -i.e. learning through cross-validation techniques- and EC methods to identify unbiased multivariate relationships -i.e. avoiding consistency issues through residual and coefficient tests for model selection. \vskip 6pt
-
-In turn, ML feature selection algorithms can be classified into three different families: Filters, Wrappers and Embedded -depending on whether they use some classifier/response variable information or not, or how variable selection is made along with the learning process. Similarly, most EC dimensionality reduction approaches can be classified into three different groups: Exhaustive, General-to-Specific and Specific-to-General -depending on the search pattern. \vskip 6pt
-
-Despite significant improvements in recent econometric developments -like PCGIVE/Autometrics or Retina algorithms, which combine some ML and EC characteristics-, available alternatives fails to fully exploit cross-validation and model averaging capabilities. \vskip 6pt
-
-Conversely, newer ML algorithms -like Convolutional Neural Networks or Bootstrap-Based LASSO, which improve complex non-linear adjustment or model selection under regularization schemes- achieved unprecedented forecast accuracy but disregarding model interpretability and/or parameter estimation issues -i.e omitting residual and coefficient tests. \vskip 6pt
-
[…]
