### 2019-07-26 c6a31c4
Update paper.tex
