### 2019-07-26 94723c4
Update paper.tex
