+++ paper/ref.bib
@@ -4,374 +4,3 @@
   year={1961},
   publisher={Princeton University Press}
 }
-
-@book{Buhlmann2011,
-  author={Bhlmann, Peter, and Van De Geer, Sara},
-  title={Statistics for high-dimensional data: methods, theory and applications},
-  year={2011},
-  publisher={Springer-Verlag Berlin Heidelberg}
-}
-
-@conference{Yu2003,
-  author={Yu, Lei, and Liu, Huan},
-  title={Feature selection for high-dimensional data: A fast correlation-based filter solution},
-  booktitle={In Proceedings of the 20th international conference on machine learning},
-  year={2003},
-  publisher={The publisher of the proceedings},
-  pages={856--863},
-  url={https://www.aaai.org/Papers/ICML/2003/ICML03-111.pdf}
-}
-
-@article{Defernez1999,
-  title={Avoiding overfitting in the analysis of high-dimensional data with artificial neural networks (ANNs)},
-  author={Defernez, Marianne, and Kemsley, Katherine},
-  journal={Analyst},
-  volume={124},
-  number={11},
-  pages={1675--1681},
-  year={1999}
-}
-
-@incollection{Aggarwal2001,
-  title={On the surprising behavior of distance metrics in high dimensional space},
-  author={Aggarwal, Charu, and Hinneburg, Alexander and Keim Daniel},
-  booktitle={International conference on database theory},
-  publisher={Springer, Berlin, Heidelberg},
-  editor={Van den Bussche, Jan and Vianu, Victor},
-  series={Lecture Notes in Computer Science, vol 1973},
-  pages={420--434},
-  year={2001}
-}
-
-@book{Bolon2016,
-  author={Bolón-Canedo, Verónica, and Sánchez-Maroño, Noelia, and Alonso-Betanzos, Amparo},
-  title={Feature selection for high-dimensional data},
-  series={ Progress in Artificial Intelligence},
-  year={2016},
-  publisher={Springer-Verlag Berlin Heidelberg}
-}
-
-@article{gluzmann2015,
-  title={Global search regression: A new automatic model-selection technique for cross-section, time-series, and panel-data regressions},
-  author={Glüzmann, Pablo, and Panigo, Demian},
-  journal={The Stata journal},
-  volume={15},
-  number={2},
-  pages={325--349},
-  year={2015}
-} 
-
-@inproceedings{Doornik09autometrics,
-  author={Doornik, Jurgen A.},
-  title={Autometrics},
-  booktitle={in Honour of David F. Hendry},
-  year={2009},
-  pages={88--121},
-  publisher={University Press}
-}
-
-@article{varian2014,
-  Author={Varian, Hal R.},
-  Title={Big Data: New Tricks for Econometrics},
-  Journal={Journal of Economic Perspectives},
-  Volume={28},
-  Number={2},
-  Year={2014},
-  Month={May},
-  Pages={3--28},
-  DOI={10.1257/jep.28.2.3},
-  URL={http://www.aeaweb.org/articles?id=10.1257/jep.28.2.3}
-}
-
-@article{perezamaral2003,
-  author={Perez-Amaral, Teodosio and Gallo, Giampiero M. and White, Halbert},
-  title={A Flexible Tool for Model Building: the Relevant Transformation of the Inputs Network Approach (RETINA)*},
-  journal={Oxford Bulletin of Economics and Statistics},
-  volume={65},
-  number={s1},
-  pages={821--838},
-  keywords={C52, C53, C45},
-  doi={10.1046/j.0305-9049.2003.00096.x},
-  url={https://onlinelibrary.wiley.com/doi/abs/10.1046/j.0305-9049.2003.00096.x},
[…]
