+++ paper/ref.bib
@@ -4,3 +4,13 @@
   year={1961},
   publisher={Princeton University Press}
 }
+
+@conference{Yu2003,
+  author={Yu, Lei, and Liu, Huan},
+  title={Feature selection for high-dimensional data: A fast correlation-based filter solution},
+  booktitle={In Proceedings of the 20th international conference on machine learning},
+  year={2003},
+  publisher={The publisher of the proceedings},
+  pages={856--863},
+  url={https://www.aaai.org/Papers/ICML/2003/ICML03-111.pdf}
+}
