+++ paper/ref.bib
@@ -5,6 +5,13 @@
   publisher={Princeton University Press}
 }
 
+@book{Buhlmann2011,
+  author={Bhlmann, Peter, and Van De Geer, Sara},
+  title={Statistics for high-dimensional data: methods, theory and applications},
+  year={2011},
+  publisher={Springer-Verlag Berlin Heidelberg}
+}
+
 @conference{Yu2003,
   author={Yu, Lei, and Liu, Huan},
   title={Feature selection for high-dimensional data: A fast correlation-based filter solution},
@@ -14,3 +21,5 @@
   pages={856--863},
   url={https://www.aaai.org/Papers/ICML/2003/ICML03-111.pdf}
 }
+
+
