+++ paper/ref.bib
@@ -22,4 +22,12 @@
   url={https://www.aaai.org/Papers/ICML/2003/ICML03-111.pdf}
 }
 
-
+@article{Defernez1999,
+  title={Avoiding overfitting in the analysis of high-dimensional data with artificial neural networks (ANNs)},
+  author={Defernez, Marianne, and Kemsley, Katherine},
+  journal={Analyst},
+  volume={124},
+  number={11},
+  pages={1675--1681},
+  year={1999}
+}
