+++ paper/ref.bib
@@ -31,3 +31,15 @@
   pages={1675--1681},
   year={1999}
 }
+
+@incollection{Aggarwal2001,
+  title={On the surprising behavior of distance metrics in high dimensional space},
+  author={Aggarwal, Charu, and Hinneburg, Alexander and Keim Daniel},
+  booktitle={International conference on database theory},
+  publisher={Springer, Berlin, Heidelberg},
+  editor={Van den Bussche, Jan and Vianu, Victor},
+  series={Lecture Notes in Computer Science, vol 1973},
+  pages={420--434},
+  year={2001}
+}
+
