+++ paper/ref.bib
@@ -43,3 +43,11 @@
   year={2001}
 }
 
+@book{Bolon2016,
+  author={Bolón-Canedo, Verónica, and Sánchez-Maroño, Noelia, and Alonso-Betanzos, Amparo},
+  title={Feature selection for high-dimensional data},
+  series={ Progress in Artificial Intelligence},
+  year={2016},
+  publisher={Springer-Verlag Berlin Heidelberg}
+}
+
