+++ paper/ref.bib
@@ -51,3 +51,13 @@
   publisher={Springer-Verlag Berlin Heidelberg}
 }
 
+@article{gluzmann2015,
+  title={Global search regression: A new automatic model-selection technique for cross-section, time-series, and panel-data regressions},
+  author={Glüzmann, Pablo, and Panigo, Demian},
+  journal={The Stata journal},
+  volume={15},
+  number={2},
+  pages={325--349},
+  year={2015}
+} 
+
