+++ paper/ref.bib
@@ -61,3 +61,12 @@
   year={2015}
 } 
 
+@inproceedings{Doornik09autometrics,
+  author={Doornik, Jurgen A.},
+  title={Autometrics},
+  booktitle={in Honour of David F. Hendry},
+  year={2009},
+  pages={88--121},
+  publisher={University Press}
+}
+
