+++ paper/ref.bib
@@ -70,3 +70,26 @@
   publisher={University Press}
 }
 
+@article{varian2014,
+  Author={Varian, Hal R.},
+  Title={Big Data: New Tricks for Econometrics},
+  Journal={Journal of Economic Perspectives},
+  Volume={28},
+  Number={2},
+  Year={2014},
+  Month={May},
+  Pages={3--28},
+  DOI={10.1257/jep.28.2.3},
+  URL={http://www.aeaweb.org/articles?id=10.1257/jep.28.2.3}
+}
+
+@article{perezamaral2003,
+  author={Perez-Amaral, Teodosio and Gallo, Giampiero M. and White, Halbert},
+  title={A Flexible Tool for Model Building: the Relevant Transformation of the Inputs Network Approach (RETINA)*},
+  journal={Oxford Bulletin of Economics and Statistics},
+  volume={65},
+  number={s1},
+  pages={821--838},
+  keywords={C52, C53, C45},
+  year={2003}
+}
