+++ paper/ref.bib
@@ -97,19 +97,11 @@
 @article{CHANDRASHEKAR201416,
   title={A survey on feature selection methods},
   journal={Computers & Electrical Engineering},
-  volume="40",
+  volume={40},
   number={1},
   pages={16--28},
   year={2014},
   issn={0045-7906},
   author={Chandrashekar, Girish and Sahin, Ferat}
-}
-
-@PhdThesis{marinucci2008,
-  author={Marinucci, Maximiliano},
-  title={Automatic prediction and model selection},
-  school={Facultad de Ciencias Económicas y Empresariales, Universidad Complutense de Madrid},
-  type={Ph.D. Thesis}
-  year={2008}
 }
 
