+++ paper/ref.bib
@@ -94,9 +94,9 @@
   year={2003}
 }
 
-@article{CHANDRASHEKAR201416,
+@article{chandras2014,
   title={A survey on feature selection methods},
-  journal={Computers & Electrical Engineering},
+  journal={Computers and Electrical Engineering},
   volume={40},
   number={1},
   pages={16--28},
