### 2019-07-26 bcd7c2b
Update paper.tex
