+++ paper/ref.bib
@@ -96,7 +96,7 @@
 
 @article{chandras2014,
   title={A survey on feature selection methods},
-  journal={Computers and Electrical Engineering},
+  journal={Computers & Electrical Engineering},
   volume={40},
   number={1},
   pages={16--28},
