+++ paper/ref.bib
@@ -105,3 +105,271 @@
   author={Chandrashekar, Girish and Sahin, Ferat}
 }
 
+
+@PhdThesis{marinucci2008,
+  author={Marinucci, Maximiliano},
+  title={Automatic prediction and model selection},
+  school={Facultad de Ciencias Económicas y Empresariales, Universidad Complutense de Madrid},
+  type={Ph.D. Thesis}
+  year={2008}
+}
+
+@article{DAVIDSON1981,
+  title={Interpreting econometric evidence: The behaviour of consumers' expenditure in the UK}
+  journal={European Economic Review},
+  volume={16},
+  number={1},
+  pages={177--192},
+  year={1981},
+  issn={0014-2921},
+  doi={https://doi.org/10.1016/0014-2921(81)90058-1},
+  url={http://www.sciencedirect.com/science/article/pii/0014292181900581},
+  author={Davidson, James E.H. and Hendry, David F.}
+}
+
+
+@article{derksen1992,
+  author={Derksen, Shelley and Keselman, H. J.},
+  title={Backward, forward and stepwise automated subset selection algorithms: Frequency of obtaining authentic and noise variables},
+  journal={British Journal of Mathematical and Statistical Psychology},
+  volume={45},
+  number={2},
+  pages={265--282},
+  doi={10.1111/j.2044-8317.1992.tb00992.x},
+  url={https://onlinelibrary.wiley.com/doi/abs/10.1111/j.2044-8317.1992.tb00992.x},
+  eprint={https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.2044-8317.1992.tb00992.x},
+  year={1992}
+}
+
+@PhdThesis{castle2006,
+  author={Castle, Jennifer},
+  title={Empirical modeling and model selection for forecasting inflation in a non-stationary world},
+  school={Nuffield College University of Oxford},
+  type={Ph.D. Thesis}
+  year={2006}
+}
+
+@article{herwartz2010,
+  author={Herwartz, Helmut},
+  title={A note on model selection in (time series) regression models – general-to-specific or specific-to-general?},
+  journal={Applied Economics Letters},
+  volume={17},
+  number={12},
+  pages={1157--1160},
+  year={2010},
+  publisher={Routledge},
+  doi={10.1080/00036840902845418},
+  URL={https://doi.org/10.1080/00036840902845418},
+  eprint={https://doi.org/10.1080/00036840902845418}
+}
+
+@article{Bzdok2018,
+  author={Bzdok, Danilo, and  Altman, Naomi, and  Krzywinski, Martin},
+  title={Statistics versus machine learning},
+  journal={Nature methods},
+  volume={15},
+  number={4},
+  pages={233--234},
+  year={2010},
+  publisher={Routledge},
+  doi={doi:10.1038/nmeth.4642}
+}
+
+@article{hoetening1999,
+  ISSN={08834237},
+  URL={http://www.jstor.org/stable/2676803},
+  author={Hoeting, Jennifer A. and Madigan, David, and Raftery, Adrian E., and Volinsky, Chris T.},
+  journal={Statistical Science},
+  number={4},
+  pages={382--401},
+  publisher={Institute of Mathematical Statistics},
+  title={Bayesian Model Averaging: A Tutorial},
+  volume={14},
+  year={1999}
+}
+
+@article{abramovich2006,
+  ISSN={09727671},
+  URL={http://www.jstor.org/stable/25053511},
+  author={Abramovich, Felix and Angelini, Claudia},
+  journal={Sankhyā: The Indian Journal of Statistics (2003-2007)},
+  number={3},
+  pages={436--460},
+  publisher={Springer},
+  title={Bayesian Maximum a posteriori Multiple Testing Procedure},
+  volume={68},
+  year={2006}
+}
[…]
