+++ paper/paper.tex
@@ -42,6 +42,86 @@ links/53eed18a0cf23733e812c10d/Global-Search-Regression-A-New-Automatic-Model-se
 origin=publication_detail}{GSREG} Stata alternatives), this Julia algorithm allows researchers to obtain the best model specification among all possible 
 covariate/feature combinations - in terms of user defined information criteria-, but up to 3165 times faster than Stata and 197 times faster than R. \vskip 6pt
 
+\section{Package's main features}
+\label{sec:packagefeatures}
+
+Written in Julia, GlobalSearchRegression is a parallel (and improved) version of the Stata-GSREG all-subset regression command (get the original code \href{https://ideas.repec.org/c/boc/bocode/s457737.html}{here}). The package structure is quite simple, as shown in figure 1:
+
+
+%%%%%%%%%%%%%%%%%% Figure 1
+\begin{figure}[h]
+\centerline{\includegraphics[width=10cm]{flowchart.png}}
+\caption{ GlobalSearchRegression.jl Structure Flowchart}
+	\label{fig:sample_figure}
+\end{figure}
+%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+Through the gsreg function of the \verb interface.jl  internal package, users set the appropriate database to be used, the general unrestricted model -GUM, which defines the search space- and additional options for model selection. With this information and complementary supporting functions and definitions provided by \verb strings.jl  -i.e. error messages-, \verb utils.jl  -i.e. equation formatting, combinatorial analysis, database manipulation, sorting results, etc.- and \verb gsreg_results.jl  -i.e. the structure to save estimation results-, the \verb core.jl  package perform the all-subset-regression algorithm explained in the pseudocode below, to obtain the following outputs: 1) a matrix -optionally exported to a csv file through \verb utils.jl-  including egression coefficients, selection criteria, observations and (optionally) t-test, residual tests, averaging-weights and out-of-sample metrics for every lternative model; 2) a text file -also displayed on screen- which contains the best model specification and (optionally) model averaging results in terms of he user-selected information criteria -see multiple examples in \verb runtest.jl-. \vskip 6pt
+\break
+
+%%%%%%%%%%%%%%%%%%%%%%% Algorithm 1
+\begin{listing}
+\hline
+\center{ \textbf{Pseudo-code for module Core.jl}}
+\vspace{2mm}
+\hline
+  \begin{enumerate}
+    
+    \item[1.] \textit{\textbf{function} gsreg\_single\_proc\_results!(args…)}
+    \begin{enumerate}
+        \item[1.1.] Select regression covariates
+        \item[1.2.] Perform QR Decomposition
+        \item[1.3.] Estimate regression results
+        \item[1.4.] Calculate additional tests
+        \item[1.5.] Save individual results in a Shared Array
+    \end{enumerate}
+    \item[] \textbf{end} \textit{function gsreg\_single\_proc\_results!(args…)}
+
+\vspace{2mm}
+    
+    \item[2.] \textit{\textbf{function} gsreg\_proc\_results!(args…)}
+    \begin{enumerate}
+        \item[2.1.] \textbf{for} $j=1$ : $((k^2-1)/ nworkers)$ \textit{\# where nworkers is the user defined number of cores, and k is the numbers of ovariates %in the GUM}
+        \begin{enumerate}
+            \item[2.1.1.] $order = (j -1) * nworkers + workerID$
+            \item[2.1.2.] \textit{gsreg\_single\_proc\_results!(order, args...)}
+        \end{enumerate}
[…]
