### 2019-07-26 3e04c66
Update paper.tex
