### 2019-07-26 8d0ce49
Update paper.tex
