### 2019-07-26 22ba4f8
Update paper.tex
