+++ paper/paper.tex
@@ -61,15 +61,6 @@ Through the gsreg function of the \verb interface.jl  internal package, users se
 \break
 
 
-\begin{algorithm}[ht]
-  \scriptsize
-  \caption{module Core.jl}
-  \begin{enumerate}
-    
-    \item[1.] \textit{\textbf{function} gsreg\_single\_proc\_results!(args…)}
- \end{enumerate}
-\end{algorithm}
-
 
 \input{bib.tex}
 \end{document}
