### 2019-07-26 d10a024
Update paper.tex
