+++ paper/paper.tex
@@ -1,7 +1,7 @@
 % JuliaCon proceedings template
 \documentclass{juliacon}
 \usepackage{amsmath}
-\usepackage{algorithm2e}
+\usepackage[linesnumbered,ruled,vlined]{algorithm2e}
 \setcounter{page}{1}
 
 \begin{document}
