### 2019-07-26 494728c
Update paper.tex
