### 2019-07-26 5bca867
Update paper.tex
