+++ paper/paper.tex
@@ -62,8 +62,7 @@ Through the gsreg function of the \verb interface.jl  internal package, users se
 
 %%%%%%%%%%%%%%%%%%%%%%% Algorithm 1
 \begin{algorithm}[ht]
-  %\scriptsize
-  \caption{module Core.jl}
+sarasa
 \end{algorithm}
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
