### 2019-07-26 3c94c2d
Update paper.tex
