### 2019-07-26 f87c3e9
Update paper.tex
