### 2019-07-26 547a705
Update paper.tex
