### 2019-07-26 82063c4
Update paper.tex
