+++ paper/paper.tex
@@ -1,8 +1,6 @@
 % JuliaCon proceedings template
 \documentclass{juliacon}
 \usepackage{amsmath}
-\usepackage{algorithm}
-\usepackage{algorithm2e}
 \setcounter{page}{1}
 
 \begin{document}
