### 2019-07-26 8d39c64
Update paper.tex
