### 2019-07-26 a23b266
Update paper.tex
