### 2019-07-26 8a3ef70
Update paper.tex
