### 2019-07-26 96d181f
Update paper.tex
