### 2019-07-26 ce39156
Update paper.tex
