### 2019-07-26 78c96e0
Update paper.tex
