### 2019-07-26 bb98b85
Update paper.tex
