### 2019-07-26 a74db8f
Update paper.tex
