### 2019-07-26 db11e84
Update paper.tex
