+++ paper/paper.tex
@@ -129,26 +129,9 @@ While some of these characteristics are shared with R and Stata alternatives (\v
 
 \subsection{OLS estimation}
 \label{subsub:ols}
+
 Efficient Ordinary Least Squares (OLS) algorithms rely on Linear Algebra operations (matrix decomposition, matrix inversion, etc.).\footnote{Optimization alternatives (i.e. Gradient descent), while quite inefficient for linear models, can also be used for non-linear estimations \cite{abdi2007}} The traditional $(X'X)^{-1}$ operation could be time-expensive with unstable solutions under certain conditions \cite{jameson1981}. A preferred method is the QR-decomposition developed by Francis \cite{francis1961} and Kublanovskaya \cite{kublanovskaya1962s}. The QR factorization allows us to decompose any full rank $N \times p$ matrix $\bar{X}$ as:
 
-\begin{equation}
-    X=QR
-\end{equation}
-
-where
-
-\begin{itemize}
-    \item[•] $Q$ is a $N \times p$ matrix with $Q'Q=I$  ; and
-    \item[•] $R$ a $p \times p$ upper triangular matrix.
-\end{itemize}
-
-The QR decomposition is fast and provides stable numerical solutions under rank-deficient matrices. Alternative factorizations are either slower (SVD decomposition) or potentially unstable (Cholesky decomposition)\cite{agarwal2014}. \vskip 6pt
-
-\verb GlobalSearchRegression.jl  OLS estimation through QR-decompositions outperform existing Julia alternatives (like \verb GLM.jl)  allowing for large speed-ups compared with \verb R-lm    and \verb Stata-regress  commands. Table 3 shows execution time differences for a 200 covariates - 1000 observations - multivariate linear regression \footnote{Execution times were calculated for alternative OLS algorithms available for all-subset-regression packages in Julia,  %and Stata.}.
-\vskip 6pt
-
-It must be notice that execution times are compilation-free, because all-subset-regression algorithms must perform thousand to millions of regressions where compilation time is absent (it only affects the first regression).
-\vskip 6pt
 
 \input{bib.tex}
 \end{document}
