### 2019-07-26 bf1c19c
Update paper.tex
