+++ paper/paper.tex
@@ -136,6 +136,12 @@ Efficient Ordinary Least Squares (OLS) algorithms rely on Linear Algebra operati
     X=QR
 \end{equation}
 
+where
+
+\begin{itemize}
+    \item[•] $Q$ is a $N \times p$ matrix with $Q'Q=I$  ; and
+    \item[•] $R$ a $p \times p$ upper triangular matrix.
+\end{itemize}
 
 \input{bib.tex}
 \end{document}
