+++ paper/paper.tex
@@ -143,6 +143,8 @@ where
     \item[•] $R$ a $p \times p$ upper triangular matrix.
 \end{itemize}
 
+The QR decomposition is fast and provides stable numerical solutions under rank-deficient matrices. Alternative factorizations are either slower (SVD decomposition) or potentially unstable (Cholesky decomposition) \cite{agarwal2014}. \vskip 6pt
+
 \input{bib.tex}
 \end{document}
 
