### 2019-07-26 0170bdb
Update paper.tex
