+++ paper/paper.tex
@@ -148,6 +148,10 @@ The QR decomposition is fast and provides stable numerical solutions under rank-
 \verb GlobalSearchRegression.jl  OLS estimation through QR-decompositions outperform existing Julia alternatives (like \verb GLM.jl)  allowing for large speed-ups compared with \verb R-lm    and \verb Stata-regress  commands. Table 3 shows execution time differences for a 200 covariates - 1000 observations - multivariate linear regression \footnote{Execution times were calculated for alternative OLS algorithms available for all-subset-regression packages in Julia, and Stata.}.
 \vskip 6pt
 
+It must be notice that execution times are compilation-free, because all-subset-regression algorithms must perform thousand to millions of regressions where compilation time is absent (it only affects the first regression).
+\vskip 6pt
+
+
 \input{bib.tex}
 \end{document}
 
