### 2019-07-26 197babf
Update paper.tex
