### 2019-07-26 13d7842
Update paper.tex
