### 2019-07-26 07d1958
Update paper.yml
