+++ paper/paper.yml
@@ -1,4 +1,4 @@
-title: "GlobalSearchRegression.jl:\\ Building bridges between Machine Learning and Econometrics in Fat-Data scenarios"
+title: "GlobalSearchRegression.jl: \\ Building bridges between Machine Learning and Econometrics in Fat-Data scenarios"
 
 keywords:
   - Julia
