### 2019-07-26 205c8e7
Update paper.tex
