+++ paper/paper.tex
@@ -60,7 +60,9 @@ Through the gsreg function of the \verb interface.jl  internal package, users se
 \break
 
 \hrule
-\textbf{Algorithm - Core.jl}
+\vspace{2mm}
+\textbf{Pseudo-code: Core.jl}
+\vspace{2mm}
 \hrule
  \begin{enumerate}
     
