### 2019-07-26 88d90ac
Update paper.tex
