+++ paper/paper.tex
@@ -113,6 +113,7 @@ Through the gsreg function of the \verb interface.jl  internal package, users se
   \end{enumerate}
 
 \textbf{end} module \verb Core.jl .
+\hrule
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 % INSERT PSEUDOCODE HERE IN TABLES OR FIGURES
