### 2019-07-26 8f1d468
Update paper.tex
