+++ paper/paper.tex
@@ -113,15 +113,96 @@ Through the gsreg function of the \verb interface.jl  internal package, users se
   \end{enumerate}
 
 \textbf{end} module \verb Core.jl .
+\vspace{2mm}
 \hrule
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-% INSERT PSEUDOCODE HERE IN TABLES OR FIGURES
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
 
 \section{Comparative performance against \textit{R} and \textit{STATA}}
 \label{sec:comparativeperf}
+
+\begin{table*}[!t]
+\tabcolsep1pt
+\centering
+\tbl{Average execution time and speed-up metrics -over 5 runs-}{
+    \scalebox{0.8}{
+    \begin{tabular}{|C{1cm}|C{1cm}|C{1cm}|C{1.5cm}|C{1.5cm}|C{1.5cm}|C{1.5cm}|C{1.5cm}|C{1.5cm}|C{1.5cm}|C{1.5cm}|}
+    \hline
+    \multirow{Cov.} & \multirow{Obs} & \multirow{Threads} & \multicolumn{3}{c|}{Execution time (in seconds)} & \multicolumn{3}{c|}{Parallel Speed-up} & \multicolumn{2}{c|}{Julia Speed-up} \\
+\cline{4-11}      &   &   & Julia & R & Stata & Julia & R & Stata & Over R & Over Stata \\
+    \hline
+    15 & 100 & 1 & 14.0 & 69.1 & 646.6 & 1.0 & 1.0 & 1.0 & 4.9 & 46.2 \\
+    \hline
+    15 & 100 & 4 & 17.0 & 33.4 & 80.3 & 0.8 & 2.1 & 8.0 & 2.0 & 4.7 \\
+    \hline
+    15 & 100 & 16 & 22.0 & 23.6 & 22.1 & 0.6 & 2.9 & 29.2 & 1.1 & 1.0 \\
+    \hline
+    15 & 100 & 32 & 31.4 & 22.3 & 15.1 & 0.4 & 3.1 & 43.0 & 0.7 & 0.5 \\
+    \hline
+    15 & 1000 & 1 & 18.6 & 74.9 & 667.5 & 1.0 & 1.0 & 1.0 & 4.0 & 35.9 \\
+    \hline
+    15 & 1000 & 4 & 18.5 & 35.3 & 81.6 & 1.0 & 2.1 & 8.2 & 1.9 & 4.4 \\
+    \hline
+    15 & 1000 & 16 & 22.2 & 24.6 & 23.3 & 0.8 & 3.1 & 28.7 & 1.1 & 1.0 \\
+    \hline
+    15 & 1000 & 32 & 31.5 & 21.5 & 16.0 & 0.6 & 3.5 & 41.6 & 0.7 & 0.5 \\
+    \hline
+    15 & 10000 & 1 & 61.6 & 142.0 & 809.0 & 1.0 & 1.0 & 1.0 & 2.3 & 13.1 \\
+    \hline
+    15 & 10000 & 4 & 30.5 & 61.4 & 114.7 & 2.0 & 2.3 & 7.1 & 2.0 & 3.8 \\
+    \hline
+    15 & 10000 & 16 & 25.7 & 38.4 & 37.3 & 2.4 & 3.7 & 21.7 & 1.5 & 1.5 \\
+    \hline
+    15 & 10000 & 32 & 33.4 & 35.5 & 26.7 & 1.8 & 4.0 & 30.3 & 1.1 & 0.8 \\
[…]
