### 2019-07-26 40452a1
Update paper.tex
