+++ paper/paper.tex
@@ -120,13 +120,12 @@ Through the gsreg function of the \verb interface.jl  internal package, users se
 \label{sec:comparativeperf}
 
 \begin{table*}[!t]
-\tabcolsep1pt
 \centering
 \tbl{Average execution time and speed-up metrics -over 5 runs-}{
     \scalebox{0.8}{
     \begin{tabular}{|C{1cm}|C{1cm}|C{1cm}|C{1.5cm}|C{1.5cm}|C{1.5cm}|C{1.5cm}|C{1.5cm}|C{1.5cm}|C{1.5cm}|C{1.5cm}|}
     \hline
-    \multirow{Cov.} & \multirow{Obs} & \multirow{Threads} & \multicolumn{3}{c|}{Execution time (in seconds)} & \multicolumn{3}{c|}{Parallel Speed-up} & \multicolumn{2}{c|}{Julia Speed-up} \\
+    Cov. & Obs & Threads & \multicolumn{3}{c|}{Execution time (in seconds)} & \multicolumn{3}{c|}{Parallel Speed-up} & \multicolumn{2}{c|}{Julia Speed-up} \\
 \cline{4-11}      &   &   & Julia & R & Stata & Julia & R & Stata & Over R & Over Stata \\
     \hline
     15 & 100 & 1 & 14.0 & 69.1 & 646.6 & 1.0 & 1.0 & 1.0 & 4.9 & 46.2 \\
