### 2019-07-26 151b2fb
Update paper.tex
