### 2019-07-26 39e5570
Update paper.tex
