### 2019-07-26 8a2da84
Update paper.tex
