### 2019-07-26 c0f55ff
Update paper.tex
