+++ paper/paper.tex
@@ -315,7 +315,7 @@ It must be notice that execution times are compilation-free, because all-subset-
 
 
 %%%%%%%%%%%%%%%%%% Table 3
-\begin{table}[h]
+\begin{table}[htbp]
   \centering
 \tbl{Execution times for different OLS algorithms in R, Stata and Julia}{
     \begin{tabular}{|c|c|c|c|c|}
