### 2019-07-26 005d63e
Update paper.tex
