### 2019-07-26 e069f01
Update paper.tex
