### 2019-07-26 1374b31
Update paper.tex
