### 2019-07-26 85b28ed
Update paper.tex
