### 2019-07-26 dea3666
Update paper.tex
