### 2019-07-26 6aac0ec
Update paper.tex
