+++ paper/paper.tex
@@ -314,8 +314,8 @@ Efficient Ordinary Least Squares (OLS) algorithms rely on Linear Algebra operati
 where
 
 \begin{itemize}
-    \item[•] $Q$ is a $N \times p$ matrix with $Q'Q=I$  ; and
-    \item[•] $R$ a $p \times p$ upper triangular matrix.
+    \item[] $Q$ is a $N \times p$ matrix with $Q'Q=I$  ; and
+    \item[] $R$ a $p \times p$ upper triangular matrix.
 \end{itemize}
 
 The QR decomposition is fast and provides stable numerical solutions under rank-deficient matrices. Alternative factorizations are either slower (SVD decomposition) or potentially unstable (Cholesky decomposition) \cite{agarwal2014}. \vskip 6pt
