### 2019-07-26 aeb6e71
Update paper.yml
