### 2019-08-14 #40 @dpanigo
Yes Valentin....
It's the final version
https://github.com/JuliaCon/proceedings-papers/blob/jcon.00040/jcon.00040/10.21105.jcon.00040.pdf

Whedon finally succed to compile it
It's something wrong?

El mié., 14 ago. 2019 a las 12:26, Valentin Churavy (<
notifications@github.com>) escribió:

> Is this the correct and up-to-date version?
>
