### 2019-08-23 #40 @dpanigo
Valentin...
Apologies for the delay  I'm in Bolivia, giving some conferences about the
Argentinean collpase (another one).
Yes I Know some qualified econometricians to contact if you want.
Please let me know how to proceed
Thanks a lot
Demian

El mar., 20 ago. 2019 a las 6:05, Valentin Churavy (<
notifications@github.com>) escribió:

> @dpanigo <https://github.com/dpanigo> do you have suggestions for
> reviewers?
>
