### 2019-09-23 #40 @Nosferican
Sure, I saw the presentation at JuliaCon. What's the timeline associated with it?
