### 2019-09-23 #53 @whedon
[REVIEW]: GlobalSearchRegression.jl: Building bridges between Machine Learning and Econometrics in Fat-Data scenarios
**Submitting author:** @dpanigo (<a href="http://orcid.org/0000-0003-2468-1632">Demian Tupac PANIGO</a>)
**Repository:** <a href="https://github.com/ParallelGSReg/JuliaCon2019" target ="_blank">https://github.com/ParallelGSReg/JuliaCon2019</a>
