### 2019-09-30 #53 @Nosferican
On the paper,
  - Authorship
    - I believe the name in the authors for the paper would be Demian Panigo rather than Panigo Demian as every other author has first name and then surname.
    - I see that the commit data shows direct contributions to the software for every author except Pablo Gluzmann and Esteban Mocskos. Did these authors contributed made major contributions to the software (even if not direct commits to the related codebase)?
  - For the methodology the reproducible appendices are broken/missing.

On the software
  - Repository is a paper standalone, but directs to relevant repositories. This should be good, if so I can check the box.
  - Repository has an MIT "Expat" license. Consider using OSI-standard text for machine detectable license.
  - Repositories still carry legacy REQUIRE files, but support is for Julia v1. REQUIRE should be purge in order to use only the Project.toml
  - CI script for Travis and code coverage can be updated. The code coverage is not displayed in the README.
  - Community guidelines are not present.
  - Documentation seems incomplete (e.g., placeholders).
  - Not entirely sure how categorical variables or rank deficient is being handled.
  - I had a few issues using the software and ran into what I believe to be bugs. Should I open issues with minimal reproducible examples at the main repository?
