### 2019-10-24 #53 @Nosferican
I believe all my comments have been addressed.
- [ ] I will check the performance/benchmark aspect to check that box (on me).
- [ ] For the community guidelines, as GitHub public repository, there is some implied support/contributing guideline / seek support, but would be nice to have something on it... For example,  issue templates / `CONTRIBUTING.md`, etc. (response solicited)
