### 2019-10-26 #53 @dpanigo
Thanks a lot José!

El jue., 24 oct. 2019 a las 14:08, José Bayoán Santiago Calderón (<
notifications@github.com>) escribió:

> I believe all my comments have been addressed.
>
>    - I will check the performance/benchmark aspect to check that box (on
>    me).
>    - For the community guidelines, as GitHub public repository, there is
>    some implied support/contributing guideline / seek support, but would be
>    nice to have something on it... For example, issue templates /
>    CONTRIBUTING.md, etc. (response solicited)
>
