### 2019-11-12 #53 @dpanigo
@andreasnoack , @Nosferican:
We are pleased to inform you that a fully functional GUI interface for GlobalSearchRegression.jl is already available at https://github.com/ParallelGSReg/GlobalSearchRegressionGUI.jl.  Documentation is under development but usage is quite simple:

- Installation: pkg> add  GlobalSearchRegressionGUI;

- Loading: julia> using GlobalSearchRegresionGUI;

- Launching: julia> gui()

P.S.: Note that the addprocs(#) command (after "using Distributed") must be defined before loading GlobalSearchRegressionGUI. Otherwise, parallel capabilities will not be available.

-
