### 2020-07-27 #53 @dpanigo
> @dpanigo thank you for your patience. Is there an archival DOI that you would like to associate with the publication?
> https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance

@vchuravy the archival DOI should link the article with the publication repo itself or with the underlying software (GlobalSearchRegression.jl?)
