### 2020-07-27 #53 @dpanigo
> @vchuravy the archival DOI should link the article with the publication repo itself or with the underlying software (GlobalSearchRegression.jl?)

Because the software is on a different repo
