+++ paper/paper.tex
@@ -216,7 +216,7 @@ A detailed speed-up analysis is also available for the 25-covariate case. In fig
 \begin{figure*}[t]
  \centerline{\includegraphics[width=16cm]{kernels.png}} 
     \vspace{2mm}
-  \caption{Execution time, speed-up and speed-up efficiency for the 25-covariate case using GlobalSearchRegression.jl}
+  \caption{Execution time, speed-up and speed-up efficiency for the 20-covariate case using GlobalSearchRegression.jl}
 \end{figure*}
 
 \begin{figure*}[t]
