### 2020-07-28 #53 @vchuravy
@arfon @matbesancon any idea why the generate pdf succeeds, but the deposit fails?

@dpanigo can you fix https://github.com/JuliaCon/proceedings-review/issues/53#issuecomment-664914256 by adding a date field?
