+++ paper/ref.bib
@@ -110,7 +110,7 @@
   author={Marinucci, Maximiliano},
   title={Automatic prediction and model selection},
   school={Facultad de Ciencias Económicas y Empresariales, Universidad Complutense de Madrid},
-  type={Ph.D. Thesis}
+  type={Ph.D. Thesis},
   year={2008}
 }
 
@@ -145,7 +145,7 @@
   author={Castle, Jennifer},
   title={Empirical modeling and model selection for forecasting inflation in a non-stationary world},
   school={Nuffield College University of Oxford},
-  type={Ph.D. Thesis}
+  type={Ph.D. Thesis},
   year={2006}
 }
 
