### 2020-07-28 #53 @dpanigo
> @arfon @matbesancon any idea why the generate pdf succeeds, but the deposit fails?
>
> @dpanigo can you fix [#53 (comment)](https://github.com/JuliaCon/proceedings-review/issues/53#issuecomment-664914256) by adding a date field?

@vchuravy the date field was present but I found a missing "comma". Fix-it. Please try again.
