+++ paper/ref.bib
@@ -115,6 +115,7 @@
 }
 
 @article{DAVIDSON1981,
+  author={Davidson, James E.H. and Hendry, David F.},
   title={Interpreting econometric evidence: The behaviour of consumers' expenditure in the UK}
   journal={European Economic Review},
   volume={16},
@@ -123,10 +124,8 @@
   year={1981},
   issn={0014-2921},
   doi={https://doi.org/10.1016/0014-2921(81)90058-1},
-  url={http://www.sciencedirect.com/science/article/pii/0014292181900581},
-  author={Davidson, James E.H. and Hendry, David F.}
+  url={http://www.sciencedirect.com/science/article/pii/0014292181900581}
 }
-
 
 @article{derksen1992,
   author={Derksen, Shelley and Keselman, H. J.},
