+++ paper/ref.bib
@@ -116,7 +116,7 @@
 
 @article{DAVIDSON1981,
   author={Davidson, James E.H. and Hendry, David F.},
-  title={Interpreting econometric evidence: The behaviour of consumers' expenditure in the UK}
+  title={Interpreting econometric evidence: The behaviour of consumers' expenditure in the UK},
   journal={European Economic Review},
   volume={16},
   number={1},
