+++ paper/ref.bib
@@ -85,7 +85,7 @@
 
 @article{perezamaral2003,
   author={Perez-Amaral, Teodosio and Gallo, Giampiero M. and White, Halbert},
-  title={A Flexible Tool for Model Building: the Relevant Transformation of the Inputs Network Approach (RETINA)*},
+  title={A Flexible Tool for Model Building: the Relevant Transformation of the Inputs Network Approach (RETINA)},
   journal={Oxford Bulletin of Economics and Statistics},
   volume={65},
   number={s1},
