### 2020-07-28 55c4551
Update paper.yml
