+++ paper/paper.yml
@@ -11,7 +11,7 @@ keywords:
 authors:
   - name: Panigo Demian
     orcid: 0000-0003-2468-1632
-    affiliation: 1, 3, 4
+    affiliation: 1, 3, 6
   - name: Pablo Glüzmann
     orcid: 0000-0001-6570-383X
     affiliation: 2
@@ -20,25 +20,27 @@ authors:
     affiliation: 1, 5
   - name: Adan Mauri Ungaro
     orcid: 0000-0003-3106-3912
-    affiliation: 3
+    affiliation: 4
   - name: Valentin Mari
     orcid: 0000-0003-1489-6525
-    affiliation: 3
+    affiliation: 4
   - name: Nicolás Monzón
     orcid: 0000-0002-4846-7916
-    affiliation: 3, 4
+    affiliation: 3, 6
   
 affiliations:
   - name: National Scientific and Technical Research Council (CONICET) 
     index: 1
-  - name: Center For Distributive, Labor and Social Studies (CEDLAS) 
+  - name: Center For Distributive, Labor and Social Studies (CEDLAS-UNLP) 
     index: 2
-  - name: Nacional University of La Plata
+  - name: Laboratory of Economics, Technology, Innovation and Finance (LETIF-UNLP)
     index: 3
-  - name: National University of Avellaneda
+  - name: National University of La Plata
     index: 4
   - name: National University of Buenos Aires
     index: 5
+  - name: National University of Avellaneda
+    index: 6
 
 date: 7 Jul 2019
 bibliography: ref.bib
