### 2020-07-28 3a64b3e
Update paper.yml
