+++ paper/paper.yml
@@ -14,7 +14,7 @@ authors:
     affiliation: 1, 3, 6
   - name: Pablo Glüzmann
     orcid: 0000-0001-6570-383X
-    affiliation: 2
+    affiliation: 1, 2
   - name: Esteban Mocskos 
     orcid: 0000-0002-6473-7672
     affiliation: 1, 5
