+++ paper/paper.yml
@@ -11,7 +11,7 @@ keywords:
 authors:
   - name: Panigo Demian
     orcid: 0000-0003-2468-1632
-    affiliation: 1, 3, 6
+    affiliation: 1, 3
   - name: Pablo Glüzmann
     orcid: 0000-0001-6570-383X
     affiliation: 1, 2
@@ -26,7 +26,7 @@ authors:
     affiliation: 4
   - name: Nicolás Monzón
     orcid: 0000-0002-4846-7916
-    affiliation: 3, 6
+    affiliation: 3
   
 affiliations:
   - name: National Scientific and Technical Research Council (CONICET) 
@@ -39,8 +39,6 @@ affiliations:
     index: 4
   - name: National University of Buenos Aires
     index: 5
-  - name: National University of Avellaneda
-    index: 6
 
 date: 7 Jul 2019
 bibliography: ref.bib
