### 2020-07-28 #53 @dpanigo
@vchuravy I´ve tried many times and I think the bug is fixed. Otherwise, let me know how to help you. Thanks
