### 2020-07-28 #53 @vchuravy
@dpanigo can you fix the missing and invalid DOIs?
