+++ paper/ref.bib
@@ -29,6 +29,7 @@
   volume={124},
   number={11},
   pages={1675--1681},
+  DOI={10.1039/A905556H},
   year={1999}
 }
 
@@ -40,6 +41,7 @@
   editor={Van den Bussche, Jan and Vianu, Victor},
   series={Lecture Notes in Computer Science, vol 1973},
   pages={420--434},
+  DOI={10.1007/3-540-44503-x_27},
   year={2001}
 }
 
@@ -58,6 +60,7 @@
   volume={15},
   number={2},
   pages={325--349},
+  DOI={10.1177/1536867x1501500201},
   year={2015}
 } 
 
@@ -67,6 +70,7 @@
   booktitle={in Honour of David F. Hendry},
   year={2009},
   pages={88--121},
+  DOI={10.1093/acprof:oso/9780199237197.003.0004},
   publisher={University Press}
 }
 
@@ -90,6 +94,7 @@
   volume={65},
   number={s1},
   pages={821--838},
+  DOI={10.1046/j.0305-9049.2003.00096.x},
   keywords={C52, C53, C45},
   year={2003}
 }
@@ -123,7 +128,7 @@
   pages={177--192},
   year={1981},
   issn={0014-2921},
-  doi={https://doi.org/10.1016/0014-2921(81)90058-1},
+  doi={10.1016/0014-2921(81)90058-1},
   url={http://www.sciencedirect.com/science/article/pii/0014292181900581}
 }
 
@@ -231,6 +236,7 @@
   volume={61},
   number={3},
   pages={401--426},
+  DOI={10.1007/3-540-46502-2_8},
   year={2001},
   publisher={Elsevier}
 }
@@ -242,6 +248,7 @@
   volume={5},
   number={1},
   pages={1--104},
+  DOI={10.1561/1900000036},
   year={2013},
   publisher={Now Publishers, Inc.}
 }
@@ -253,6 +260,7 @@
   volume={6},
   number={3},
   pages={536--565},
+  DOI={10.1080/15598608.2012.695705},
   year={2012},
   publisher={Springer}
 }
@@ -263,7 +271,8 @@
   journal={IEEE Concurrency},
   number={3},
   pages={27--36},
-  year={1994},
+  DOI={10.2172/205045},
+  year={1996},
   publisher={IEEE}
 }
 
@@ -277,7 +286,7 @@
   year={2017},
   series={Emerging Trends in Computer Science and Applied Computing},
   isbn={978-0-12-803738-6},
-  doi={https://doi.org/10.1016/B978-0-12-803738-6.00006-9},
+  doi={10.1016/b978-0-12-803738-6.00006-9},
   url={http://www.sciencedirect.com/science/article/pii/B9780128037386000069},
   author={N. Moreano and A.C.M.A.de Melo}
 }
@@ -309,6 +318,7 @@
   volume={34},
   number={5},
   pages={151--162},
+  DOI={10.1145/1168918.1168877}
   year={2006},
   publisher={ACM}
 }
@@ -320,6 +330,7 @@
   volume={4},
   number={3},
   pages={265--271},
+  DOI={10.1093/comjnl/4.3.265},
   year={1961},
   publisher={Oxford University Press}
 }
@@ -331,6 +342,7 @@
   volume={1},
   number={3},
   pages={637--657},
+  DOI={10.1016/0041-5553(63)90168-X},
   year={1962},
   publisher={Elsevier}
 }
@@ -359,6 +371,7 @@
   title={The method of least squares},
   author={Abdi, Herv{\'e}},
   journal={Encyclopedia of Measurement and Statistics. CA, USA: Thousand Oaks},
+  DOI={10.1007/3-540-31720-1_2},
   year={2007}
 }
 
