+++ paper/ref.bib
@@ -312,15 +312,40 @@
 }
 
 @article{gordon2006,
-  title={Exploiting coarse-grained task, data, and pipeline parallelism in stream programs},
-  author={Gordon, Michael I and Thies, William and Amarasinghe, Saman},
-  journal={ACM SIGARCH Computer Architecture News},
-  volume={34},
-  number={5},
-  pages={151--162},
-  DOI={10.1145/1168918.1168877}
-  year={2006},
-  publisher={ACM}
+  author = {Gordon, Michael I. and Thies, William and Amarasinghe, Saman},
+  title = {Exploiting Coarse-Grained Task, Data, and Pipeline Parallelism in Stream Programs},
+  year = {2006},
+  issue_date = {November 2006},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  volume = {41},
+  number = {11},
+  issn = {0362-1340},
+  url = {https://doi.org/10.1145/1168918.1168877},
+  DOI = {10.1145/1168918.1168877},
+  journal = {SIGPLAN Not.},
+  month = {oct},
+  pages = {151–162},
+  numpages = {12},
+  keywords = {multicore, software pipelining, Raw, coarse-grained dataflow, StreamIt, streams}
+}
+
+
+@inproceedings{10.1145/1168857.1168877,
+author = {Gordon, Michael I. and Thies, William and Amarasinghe, Saman},
+title = {Exploiting Coarse-Grained Task, Data, and Pipeline Parallelism in Stream Programs},
+year = {2006},
+isbn = {1595934510},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/1168857.1168877},
+doi = {10.1145/1168857.1168877},
+booktitle = {Proceedings of the 12th International Conference on Architectural Support for Programming Languages and Operating Systems},
+pages = {151–162},
+numpages = {12},
+keywords = {streams, Raw, coarse-grained dataflow, StreamIt, software pipelining, multicore},
+location = {San Jose, California, USA},
+series = {ASPLOS XII}
 }
 
 @article{francis1961,
