### 2020-07-30 #53 @nicomzn
@vchuravy when we try to enter https://doi.org/10.21105/jcon.00053, this message results

![image](https://user-images.githubusercontent.com/52471741/88867273-66298580-d1e3-11ea-9330-c8d2c9396d3f.png)

do you have to do something else?
