### 2020-07-30 #53 @arfon
@nicomzn - this should be fixed now. There was a temporary issue with Crossref.
