### 2020-07-30 #53 @nicomzn
@arfon there is a problem with the first author: It's Demian Panigo, no Panigo Demian, and the citation would be Panigo et al., (2020). Can we fix it modifying the underlying file?
