### 2020-07-30 #53 @Nosferican
I had the same issue for Econometrics.jl as my name shows up as Calderón, José Bayoán Santiago instead of Santiago Calderón, José Bayoán. Ideally that would be fixed by the bib file / metadata using `surname, given name` rather than `first middle last name` and inferring through that heuristic.
