### 2020-07-30 #53 @nicomzn
@Nosferican thanks José. Maybe we could modify the paper.yml file from the repository and run whedon again. Is it possible? @vchuravy
