### 2020-07-30 #53 @arfon
@vchuravy @matbesancon - we probably need some better guidance for authors here? https://github.com/ParallelGSReg/JuliaCon2019/blob/master/paper/paper.yml#L12
