+++ paper/paper.yml
@@ -9,7 +9,7 @@ keywords:
   - Fat-Data
 
 authors:
-  - name: Panigo Demian
+  - name: Demian Panigo
     orcid: 0000-0003-2468-1632
     affiliation: 1, 3
   - name: Pablo Glüzmann
