### 2020-07-30 #53 @arfon
@nicomzn @Nosferican - both of your papers should be fixed now.

@nicomzn - yours might take a little while to show up as updated on the website (the PDF) as there is caching in place.
