### 2020-07-31 #53 @nicomzn
@arfon @Nosferican Thank you all for the help, the last question: is the headers' date correct? For each documents is "1(1), 2019", maybe it's a bug.
