### 2020-08-07 #53 @arfon
Where is the header sorry?

On Fri, Jul 31, 2020 at 09:43, Nicolas Monzon <notifications@github.com>
wrote:

>
>
> @arfon <https://github.com/arfon> @Nosferican
> <https://github.com/Nosferican> Thank you all for the help, the last
> question: is the headers' date correct? For each documents is "1(1), 2019",
> maybe it's a bug.
>
>
>
>
