### 2020-08-11 #53 @arfon
> @arfon here for example

Got it, thanks. @matbesancon @vchuravy - these are currently being passed in by the Ruby prep scripts with the paper (e.g. https://github.com/ParallelGSReg/JuliaCon2019/blob/master/paper/prep.rb#L12-L15).

It looks like we can override them with environment variables on the server side if necessary but what's the right thing to do here? This is a 2019 paper published in _2020_. I think the `1(1)` may be correct (issue 1, volume 1) but the year is probably incorrect right?
