### 2020-08-11 #53 @matbesancon
good question. Yes I think we should override this on the server with environment variables
